timer doesn't work; rest works

+++ src/motController.h
@@ -31,7 +31,7 @@
 
 void motCntrlTask(void* pv);
 void setRPM(int16_t rpm);
-int16_t* getRPMref(void);
+int32_t* getRPMref(void);
 
 typedef enum
 {
